+++ benchmarks/array/circularShift.bench.ts
@@ -1,13 +1,13 @@
 import * as _ from 'radashi'
 import { bench } from 'vitest'
 
-describe('shift', () => {
+describe('circularShift', () => {
   bench('with non-empty array', () => {
     const arr = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-    _.shift(arr, 3)
+    _.circularShift(arr, 3)
   })
 
   bench('with empty array', () => {
-    _.shift([], -3)
+    _.circularShift([], -3)
   })
 })