+++ vitest.config.ts
@@ -9,12 +9,7 @@ export default defineConfig(({ mode }) => ({
     globals: true,
     coverage: {
       thresholds: { 100: true },
-      exclude: [
-        '*.config.ts',
-        'benchmarks/**',
-        'scripts/**',
-        'tests/**/*.test-d.ts',
-      ],
+      include: ['src/**'],
     },
   },
   resolve: {